added model and corrected loss

+++ project/model/loss.py
@@ -2,4 +2,4 @@ import torch.nn.functional as F
 
 
 def nll_loss(output, target):
-    return F.nll_loss(output, target)
+    return F.nll_loss(output, target.long())
